ocr-invoice-app/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pytesseract==0.3.10
EasyOCR==1.6.2
Pillow==9.5.0
pdf2image==1.16.2
opencv-python==4.7.0.72
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pytest==7.3.2
pytest-cov==4.1.0
httpx==0.27.0
